refactor(layout): dedupe byte-identical components across TileLayout.{win32,linux,darwin}.tsx
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
NodeBackdrops, MagnifiedPaneOverlay, DisplayNodesWrapper, Placeholder, and
(confirmed 100% behaviorally identical beyond comments) OverlayNodeWrapper/
OverlayNode were independently duplicated across all three platform files.
Extracted to tilelayout-shared.tsx; the shared module-level drag state
(dragState.nodeId/layoutModel/node, isCrossTabDrag) moved to
tilelayout-drag-state.ts so all three platforms' DisplayNode implementations
(which stay platform-specific — different WebView2/WebKitGTK/WKWebView drag
registration) read/write one shared instance instead of three copies.

win32: 1015 -> 533 lines. linux: 922 -> 472. darwin: 907 -> 456.
(~1383 duplicated lines removed total.) tsc --noEmit clean; full layout test
suite (136 tests) unaffected.

<!-- agentmux:agent_id=agentx -->
