+++ frontend/layout/lib/TileLayout.linux.tsx
@@ -11,39 +11,33 @@
 // non-draggable). No explicit draggable="false" on any parent → no WebKitGTK
 // breakage, and drag is correctly restricted to the header.
 
-import { getApi, getSettingsKeyAtom } from "@/app/store/global";
-import { draggable, dropTargetForElements, monitorForElements } from "@atlaskit/pragmatic-drag-and-drop/element/adapter";
+import { getApi } from "@/app/store/global";
+import { draggable } from "@atlaskit/pragmatic-drag-and-drop/element/adapter";
 import { preventUnhandled } from "@atlaskit/pragmatic-drag-and-drop/prevent-unhandled";
 import clsx from "clsx";
 import { toPng } from "html-to-image";
-import { For, Show, createEffect, createMemo, createSignal, onCleanup, onMount } from "solid-js";
+import { createEffect, createMemo, createSignal, onCleanup, onMount } from "solid-js";
 import type { JSX } from "solid-js";
 import { Key } from "@solid-primitives/keyed";
 import { debounce, throttle } from "throttle-debounce";
 import { LayoutModel } from "./layoutModel";
 import { useNodeModel, useTileLayout } from "./layoutModelHooks";
-import { activeKeyFor } from "./layoutNodeModels";
 import "./tilelayout.scss";
-import {
-    DropDirection,
-    LayoutNode,
-    LayoutTreeActionType,
-    LayoutTreeComputeMoveNodeAction,
-    ResizeHandleProps,
-    TileLayoutContents,
-} from "./types";
-import { determineDropDirection } from "./utils";
+import { LayoutNode, LayoutTreeActionType, ResizeHandleProps, TileLayoutContents } from "./types";
 import { setCurrentDragPayload } from "@/app/drag/CrossWindowDragMonitor";
 import { setTileDragInFlight } from "./dragInFlight";
+import { clearCrossTabDrop } from "./crossTabDrag";
+import { dragState } from "./tilelayout-drag-state";
 import {
-    clampCrossTabDirection,
-    clearCrossTabDrop,
-    noteCrossTabDrop,
-    redockDraggedPane,
-    takeCrossTabDropFor,
-} from "./crossTabDrag";
+    DisplayNodesWrapper,
+    MagnifiedPaneOverlay,
+    NodeBackdrops,
+    OverlayNodeWrapper,
+    Placeholder,
+    tileItemType,
+} from "./tilelayout-shared";
 
-export const tileItemType = "TILE_ITEM";
+export { tileItemType };
 
 // Data stored in the HTML5 drag event dataTransfer
 const DRAG_DATA_KEY = "application/x-tile-node-id";
@@ -68,21 +62,6 @@ export interface TileLayoutProps {
 
 const DragPreviewWidth = 300;
 const DragPreviewHeight = 300;
-
-// Global drag state — track the node being dragged and current cursor position.
-let globalDragNodeId: string | null = null;
-let globalDragLayoutModel: LayoutModel | null = null;
-// The dragged LayoutNode itself — needed by cross-tab drags, where the
-// TARGET tab's overlay must compute a ghost placeholder for a node that
-// doesn't exist in its own tree (LayoutTreeComputeMoveNodeAction.nodeToMove).
-let globalDragNode: LayoutNode | null = null;
-
-// True when the drag in progress originated in a DIFFERENT tab's layout than
-// `model`'s — i.e. the user spring-switched tabs mid-drag
-// (SPEC_PANE_DRAG_TO_TAB_2026_07_10.md) and is now hovering this tab's panes.
-function isCrossTabDrag(model: LayoutModel): boolean {
-    return globalDragLayoutModel != null && globalDragLayoutModel !== model;
-}
 
 function TileLayoutComponent(props: TileLayoutProps) {
     const layoutModel = useTileLayout(props.tabAtom, props.contents);
@@ -141,10 +120,10 @@ function TileLayoutComponent(props: TileLayoutProps) {
     );
 
     // Handle drag-over for bounds checking to clear pending action when cursor leaves container.
-    // Guard: only fire during an active tile drag (globalDragNodeId set). Tab DnD must not
+    // Guard: only fire during an active tile drag (dragState.nodeId set). Tab DnD must not
     // trigger persistToBackend via treeReducer, which caused a crash on Linux (see drag.rs notes).
     const checkForCursorBounds = debounce(100, (x: number, y: number) => {
-        if (!globalDragNodeId) return;
+        if (!dragState.nodeId) return;
         if (layoutModel.displayContainerRef?.current) {
             const displayContainerRect = layoutModel.displayContainerRef.current.getBoundingClientRect();
             const normalizedX = x - displayContainerRect.x;
@@ -184,7 +163,7 @@ function TileLayoutComponent(props: TileLayoutProps) {
                 class="display-container"
             >
                 <ResizeHandleWrapper layoutModel={layoutModel} />
-                <DisplayNodesWrapper layoutModel={layoutModel} />
+                <DisplayNodesWrapper layoutModel={layoutModel} DisplayNode={DisplayNode} />
             </div>
 
             {/* Magnify layer — outside display-container to avoid stacking context issues */}
@@ -197,133 +176,6 @@ function TileLayoutComponent(props: TileLayoutProps) {
     );
 }
 export const TileLayout = TileLayoutComponent;
-
-function NodeBackdrops(props: { layoutModel: LayoutModel }) {
-    const blockBlurAtom = getSettingsKeyAtom("window:magnifiedblockblursecondarypx");
-    const blockBlur = () => blockBlurAtom();
-    const ephemeralNode = () => props.layoutModel.ephemeralNode();
-    const magnifiedNodeId = () => props.layoutModel.magnifiedNodeIdAtom();
-
-    const [showMagnifiedBackdrop, setShowMagnifiedBackdrop] = createSignal(!!magnifiedNodeId());
-    const [showEphemeralBackdrop, setShowEphemeralBackdrop] = createSignal(!!ephemeralNode());
-
-    const debouncedSetMagnifyBackdrop = debounce(100, () => setShowMagnifiedBackdrop(true));
-
-    createEffect(() => {
-        const mId = magnifiedNodeId();
-        const eph = ephemeralNode();
-        if (mId && !showMagnifiedBackdrop()) {
-            debouncedSetMagnifyBackdrop();
-        }
-        if (!mId) {
-            setShowMagnifiedBackdrop(false);
-        }
-        if (eph && !showEphemeralBackdrop()) {
-            setShowEphemeralBackdrop(true);
-        }
-        if (!eph) {
-            setShowEphemeralBackdrop(false);
-        }
-    });
-
-    const blockBlurStr = () => `${blockBlur()}px`;
-
-    return (
-        <>
-            <Show when={showMagnifiedBackdrop()}>
-                <div
-                    class="magnified-node-backdrop"
-                    onClick={() => {
-                        props.layoutModel.magnifyNodeToggle(magnifiedNodeId());
-                    }}
-                    style={{ "--block-blur": blockBlurStr() } as JSX.CSSProperties}
-                />
-            </Show>
-            <Show when={showEphemeralBackdrop()}>
-                <div
-                    class="ephemeral-node-backdrop"
-                    onClick={() => {
-                        props.layoutModel.closeNode(ephemeralNode()?.id);
-                    }}
-                    style={{ "--block-blur": blockBlurStr() } as JSX.CSSProperties}
-                />
-            </Show>
-        </>
-    );
-}
-
-/**
- * Renders the magnified pane in a dedicated overlay container outside the display-container,
- * bypassing CSS stacking context issues that prevent z-index from working on tile-nodes.
- */
-const MagnifiedPaneOverlay = (props: { layoutModel: LayoutModel }) => {
-    const magnifiedNodeId = () => props.layoutModel.magnifiedNodeIdAtom();
-    const magnifiedBlockSizeAtom = getSettingsKeyAtom("window:magnifiedblocksize");
-    const magnifiedNodeSize = () => magnifiedBlockSizeAtom() ?? 1.0;
-
-    // Find the leaf node matching the magnified node ID
-    const magnifiedNode = createMemo(() => {
-        const nodeId = magnifiedNodeId();
-        if (!nodeId) return null;
-        return props.layoutModel.leafs().find((leaf) => leaf.id === nodeId) ?? null;
-    });
-
-    // Escape key handler to unmagnify
-    const onKeyDown = (e: KeyboardEvent) => {
-        if (e.key === "Escape" && magnifiedNodeId()) {
-            props.layoutModel.magnifyNodeToggle(magnifiedNodeId());
-        }
-    };
-
-    onMount(() => window.addEventListener("keydown", onKeyDown));
-    onCleanup(() => window.removeEventListener("keydown", onKeyDown));
-
-    const containerStyle = createMemo(() => {
-        const size = magnifiedNodeSize();
-        const margin = ((1 - size) / 2) * 100;
-        return {
-            top: `${margin}%`,
-            left: `${margin}%`,
-            width: `${size * 100}%`,
-            height: `${size * 100}%`,
-        } as JSX.CSSProperties;
-    });
-
-    // The magnified pane is NOT rendered here. `DisplayNode` keeps the single
-    // `.tile-leaf` instance and reparents it into `.magnify-pane` below while
-    // magnified — one Block, one ViewModel, one browser-pane HWND across the
-    // magnify/restore cycle. Rendering a second copy here left the restored
-    // pane without zoom (terminal) or black (browser-pane native window
-    // destroyed on the duplicate's unmount). See
-    // SPEC_MAGNIFY_ZOOM_IMPLEMENTATION_2026-05-21.md.
-    return (
-        <Show when={magnifiedNode()}>
-            <div class="magnify-container" style={containerStyle()}>
-                <div
-                    class="magnify-pane"
-                    ref={(el) => {
-                        props.layoutModel.magnifyMount._set(el);
-                        onCleanup(() => props.layoutModel.magnifyMount._set(null));
-                    }}
-                />
-            </div>
-        </Show>
-    );
-};
-
-interface DisplayNodesWrapperProps {
-    layoutModel: LayoutModel;
-}
-
-const DisplayNodesWrapper = (props: DisplayNodesWrapperProps) => {
-    const leafs = () => props.layoutModel.leafs();
-
-    return (
-        <Key each={leafs()} by={activeKeyFor}>
-            {(node) => <DisplayNode layoutModel={props.layoutModel} node={node()} />}
-        </Key>
-    );
-};
 
 interface DisplayNodeProps {
     layoutModel: LayoutModel;
@@ -428,9 +280,9 @@ const DisplayNode = (props: DisplayNodeProps) => {
                     // preview just vanishes on release. In-window rearrange
                     // still works via pragmatic-dnd's own drop targets.
                     preventUnhandled.start();
-                    globalDragNodeId = props.node.id;
-                    globalDragLayoutModel = props.layoutModel;
-                    globalDragNode = props.node;
+                    dragState.nodeId = props.node.id;
+                    dragState.layoutModel = props.layoutModel;
+                    dragState.node = props.node;
                     setTileDragInFlight(true);
                     clearCrossTabDrop();
                     props.layoutModel.activeDrag._set(true);
@@ -444,9 +296,9 @@ const DisplayNode = (props: DisplayNodeProps) => {
                 },
                 onDrop: () => {
                     preventUnhandled.stop();
-                    globalDragNodeId = null;
-                    globalDragLayoutModel = null;
-                    globalDragNode = null;
+                    dragState.nodeId = null;
+                    dragState.layoutModel = null;
+                    dragState.node = null;
                     setTileDragInFlight(false);
                     props.layoutModel.activeDrag._set(false);
                     setIsDragging(false);
@@ -522,261 +374,6 @@ const DisplayNode = (props: DisplayNodeProps) => {
             {leafContent()}
             {previewElement()}
         </div>
-    );
-};
-
-interface OverlayNodeWrapperProps {
-    layoutModel: LayoutModel;
-    // Issue #836: effective overlay transform that holds the last in-drag
-    // value for 150ms after onDrop so .placeholder-container does not
-    // teleport off-screen before the inner .placeholder exit fade plays.
-    effectiveOverlayTransform: () => JSX.CSSProperties | undefined;
-}
-
-const OverlayNodeWrapper = (props: OverlayNodeWrapperProps) => {
-    const leafs = () => props.layoutModel.leafs();
-    const activeDrag = () => props.layoutModel.activeDrag();
-    let overlayContainerRef: HTMLDivElement | undefined;
-
-    // Overlay is always positioned at top:0 so pragmatic-dnd drop targets
-    // are registered in the correct location. pointer-events toggles between
-    // "none" (pass-through for normal clicks) and "auto" (receive drag events).
-    // activeDrag is set by pragmatic-dnd's onDragStart callback which fires
-    // AFTER the browser commits the drag, so this toggle is safe.
-    //
-    // Issue #836: the transform comes from effectiveOverlayTransform (held
-    // during the 150ms exit window) but pointer-events flips to "none"
-    // IMMEDIATELY when activeDrag goes false — we don't want the held
-    // overlay to intercept clicks during the hold.
-    const isActiveDrag = () => props.layoutModel.activeDrag();
-    const overlayStyle = createMemo<JSX.CSSProperties>(() => ({
-        ...props.effectiveOverlayTransform(),
-        top: "0px",
-        "pointer-events": isActiveDrag() ? "auto" : "none",
-    }));
-
-    // Fallback nearest-pane drop computation for dead spots between
-    // per-pane overlay-nodes. Mirrors TileLayout.win32.tsx — see that
-    // file for the design rationale.
-    const fallbackComputeDropDirection = throttle(50, (clientX: number, clientY: number) => {
-        const dragNodeId = globalDragNodeId;
-        if (!dragNodeId) return;
-        const container = props.layoutModel.displayContainerRef?.current;
-        if (!container) return;
-        const containerRect = container.getBoundingClientRect();
-        const offset = { x: clientX - containerRect.x, y: clientY - containerRect.y };
-
-        // Origin-rect guard — see TileLayout.win32.tsx for rationale.
-        const originRect = props.layoutModel.getNodeRectById(dragNodeId);
-        if (
-            originRect &&
-            offset.x >= originRect.left &&
-            offset.x <= originRect.left + originRect.width &&
-            offset.y >= originRect.top &&
-            offset.y <= originRect.top + originRect.height
-        ) {
-            props.layoutModel.treeReducer({ type: LayoutTreeActionType.ClearPendingAction });
-            return;
-        }
-
-        let bestLeafId: string | null = null;
-        let bestRect: { top: number; left: number; width: number; height: number } | null = null;
-        let bestDist = Infinity;
-        for (const leaf of props.layoutModel.leafs()) {
-            if (leaf.id === dragNodeId) continue;
-            const rect = props.layoutModel.getNodeRectById(leaf.id);
-            if (!rect) continue;
-            const cx = rect.left + rect.width / 2;
-            const cy = rect.top + rect.height / 2;
-            const dx = offset.x - cx;
-            const dy = offset.y - cy;
-            const dist = dx * dx + dy * dy;
-            if (dist < bestDist) {
-                bestDist = dist;
-                bestLeafId = leaf.id;
-                bestRect = rect;
-            }
-        }
-        if (!bestLeafId || !bestRect) return;
-
-        // Clamp the cursor offset to the chosen rect — see
-        // TileLayout.win32.tsx for full rationale. (codex P2 on PR #838.)
-        const clampedOffset = {
-            x: Math.max(bestRect.left, Math.min(bestRect.left + bestRect.width, offset.x)),
-            y: Math.max(bestRect.top, Math.min(bestRect.top + bestRect.height, offset.y)),
-        };
-
-        const crossTab = isCrossTabDrag(props.layoutModel);
-        const rawDirection = determineDropDirection(bestRect, clampedOffset);
-        // Cross-tab Outer* clamp — see clampCrossTabDirection.
-        const direction = crossTab ? clampCrossTabDirection(rawDirection) : rawDirection;
-        props.layoutModel.treeReducer({
-            type: LayoutTreeActionType.ComputeMove,
-            nodeId: bestLeafId,
-            nodeToMoveId: dragNodeId,
-            direction,
-            // Cross-tab: see OverlayNode.computeDropDirection — preview only.
-            nodeToMove: crossTab ? globalDragNode : undefined,
-        } as LayoutTreeComputeMoveNodeAction);
-        if (crossTab) {
-            const blockId = globalDragNode?.data?.blockId;
-            const sourceTabId = globalDragLayoutModel?.tabAtom()?.oid;
-            const targetTabId = props.layoutModel.tabAtom()?.oid;
-            const bestLeaf = props.layoutModel.leafs().find((l) => l.id === bestLeafId);
-            const targetBlockId = bestLeaf?.data?.blockId;
-            if (blockId && sourceTabId && targetTabId && targetBlockId
-                && direction !== undefined && direction !== DropDirection.Center) {
-                noteCrossTabDrop({ blockId, sourceTabId, targetTabId, targetBlockId, direction });
-            } else {
-                noteCrossTabDrop(null);
-            }
-        }
-    });
-
-    onMount(() => {
-        if (!overlayContainerRef) return;
-        const cleanup = dropTargetForElements({
-            element: overlayContainerRef,
-            canDrop: ({ source }) => source.data.type === tileItemType,
-            onDrag: ({ location }) => {
-                if (location.current.dropTargets[0]?.element !== overlayContainerRef) return;
-                const cursor = location.current.input;
-                fallbackComputeDropDirection(cursor.clientX, cursor.clientY);
-            },
-            onDragLeave: () => {
-                props.layoutModel.treeReducer({ type: LayoutTreeActionType.ClearPendingAction });
-            },
-            onDrop: ({ location }) => {
-                // Catches drops that landed in a dead spot. Clearing the
-                // payload is critical — without it the cross-window dragend
-                // monitor sees a still-set payload and can misread the
-                // release as a tear-off.
-                setCurrentDragPayload(null);
-                // Only act when this container is the innermost matched
-                // target — for drops on an inner overlay-node, that node's
-                // own onDrop already handled the commit/redock.
-                if (location.current.dropTargets[0]?.element !== overlayContainerRef) return;
-                const crossTabDrop = takeCrossTabDropFor(props.layoutModel.tabAtom()?.oid);
-                if (crossTabDrop) {
-                    props.layoutModel.treeReducer({ type: LayoutTreeActionType.ClearPendingAction });
-                    redockDraggedPane(crossTabDrop);
-                } else {
-                    props.layoutModel.onDrop();
-                }
-            },
-        });
-        onCleanup(cleanup);
-    });
-
-    return (
-        <div ref={overlayContainerRef} class="overlay-container" style={overlayStyle()}>
-            <Key each={leafs()} by={activeKeyFor}>
-                {(node) => <OverlayNode layoutModel={props.layoutModel} node={node()} />}
-            </Key>
-        </div>
-    );
-};
-
-interface OverlayNodeProps {
-    layoutModel: LayoutModel;
-    node: LayoutNode;
-}
-
-/**
- * An overlay representing the true flexbox layout of the LayoutTreeState.
- * Holds the drop targets for moving around nodes.
- */
-const OverlayNode = (props: OverlayNodeProps) => {
-    const nodeModel = useNodeModel(props.layoutModel, props.node);
-    const additionalProps = () => nodeModel.additionalProps();
-    let overlayRef: HTMLDivElement | undefined;
-
-    // Throttled drop-direction computation (same logic as before, used by pragmatic-dnd onDrag)
-    const computeDropDirection = throttle(50, (clientX: number, clientY: number) => {
-        const dragNodeId = globalDragNodeId;
-        if (!dragNodeId || dragNodeId === props.node.id) return;
-
-        const crossTab = isCrossTabDrag(props.layoutModel);
-        if (props.layoutModel.displayContainerRef?.current && additionalProps()?.rect) {
-            const containerRect = props.layoutModel.displayContainerRef.current.getBoundingClientRect();
-            const offset = { x: clientX - containerRect.x, y: clientY - containerRect.y };
-            const rawDirection = determineDropDirection(additionalProps().rect, offset);
-            // Cross-tab: Outer* directions are clamped to their inner
-            // equivalents so the ghost matches the committed split — see
-            // clampCrossTabDirection in crossTabDrag.ts.
-            const direction = crossTab ? clampCrossTabDirection(rawDirection) : rawDirection;
-            props.layoutModel.treeReducer({
-                type: LayoutTreeActionType.ComputeMove,
-                nodeId: props.node.id,
-                nodeToMoveId: dragNodeId,
-                direction,
-                // Cross-tab: the dragged node isn't in this tree; pass it so
-                // the ghost placeholder can still be computed (preview only —
-                // the drop below routes to redockDraggedPane, never a local commit).
-                nodeToMove: crossTab ? globalDragNode : undefined,
-            } as LayoutTreeComputeMoveNodeAction);
-            // Capture the cross-tab drop record NOW (globals are alive during
-            // onDrag) — drop handlers must not read the drag globals, since
-            // the source draggable's own onDrop may null them first.
-            if (crossTab) {
-                const blockId = globalDragNode?.data?.blockId;
-                const sourceTabId = globalDragLayoutModel?.tabAtom()?.oid;
-                const targetTabId = props.layoutModel.tabAtom()?.oid;
-                const targetBlockId = props.node.data?.blockId;
-                if (blockId && sourceTabId && targetTabId && targetBlockId
-                    && direction !== undefined && direction !== DropDirection.Center) {
-                    noteCrossTabDrop({ blockId, sourceTabId, targetTabId, targetBlockId, direction });
-                } else {
-                    noteCrossTabDrop(null);
-                }
-            }
-        } else {
-            props.layoutModel.treeReducer({
-                type: LayoutTreeActionType.ClearPendingAction,
-            });
-            if (crossTab) noteCrossTabDrop(null);
-        }
-    });
-
-    onMount(() => {
-        if (!overlayRef) return;
-        const cleanup = dropTargetForElements({
-            element: overlayRef,
-            canDrop: ({ source }) => source.data.type === tileItemType && source.data.nodeId !== props.node.id,
-            onDrag: ({ location }) => {
-                const cursor = location.current.input;
-                computeDropDirection(cursor.clientX, cursor.clientY);
-            },
-            onDragLeave: () => {
-                props.layoutModel.treeReducer({ type: LayoutTreeActionType.ClearPendingAction });
-            },
-            onDrop: () => {
-                // Valid in-window drop — clear cross-window payload so dragend monitor skips.
-                setCurrentDragPayload(null);
-                const crossTabDrop = takeCrossTabDropFor(props.layoutModel.tabAtom()?.oid);
-                if (crossTabDrop) {
-                    // Cross-tab drop: the pending Move references a node from
-                    // ANOTHER tab's tree — committing it locally would insert
-                    // a leaf this tab doesn't own (Tab.blockids unchanged →
-                    // dangling ref). Clear the preview and route through the
-                    // sanctioned redock RPC instead.
-                    props.layoutModel.treeReducer({ type: LayoutTreeActionType.ClearPendingAction });
-                    redockDraggedPane(crossTabDrop);
-                } else {
-                    props.layoutModel.onDrop();
-                }
-            },
-        });
-        onCleanup(cleanup);
-    });
-
-    return (
-        <div
-            ref={overlayRef}
-            class="overlay-node"
-            id={props.node.id}
-            style={additionalProps()?.transform as JSX.CSSProperties}
-        />
     );
 };
 
@@ -873,50 +470,3 @@ const ResizeHandle = (props: ResizeHandleComponentProps) => {
     );
 };
 
-interface PlaceholderProps {
-    layoutModel: LayoutModel;
-    style: JSX.CSSProperties;
-}
-
-/**
- * An overlay to preview pending actions on the layout tree.
- * Two-div split mirrors `TileLayout.win32.tsx`: outer `.placeholder-sizer`
- * carries the inline transform (the only thing animated during drag-move,
- * compositor-only — see `tilelayout.scss` `&.animate` selector), inner
- * `.placeholder` carries the visual appearance + enter/exit animation.
- * Delayed unmount holds the SolidJS `<Show>` node alive for 150ms so
- * the CSS `.exiting` fade-out can play before the DOM node is removed.
- */
-const Placeholder = (props: PlaceholderProps) => {
-    const placeholderTransform = () => props.layoutModel.placeholderTransform();
-    const [visible, setVisible] = createSignal(false);
-    const [exiting, setExiting] = createSignal(false);
-    const [lastTransform, setLastTransform] = createSignal<JSX.CSSProperties | null>(null);
-    let exitTimer: ReturnType<typeof setTimeout> | null = null;
-
-    createEffect(() => {
-        const xf = placeholderTransform();
-        if (xf) {
-            setLastTransform(xf as JSX.CSSProperties);
-            if (exitTimer) { clearTimeout(exitTimer); exitTimer = null; }
-            setExiting(false);
-            setVisible(true);
-        } else if (visible()) {
-            if (exitTimer) { clearTimeout(exitTimer); exitTimer = null; }
-            setExiting(true);
-            exitTimer = setTimeout(() => { setVisible(false); setExiting(false); }, 150);
-        }
-    });
-
-    onCleanup(() => { if (exitTimer) clearTimeout(exitTimer); });
-
-    return (
-        <div class="placeholder-container" style={props.style}>
-            <Show when={visible()}>
-                <div class="placeholder-sizer" style={lastTransform() ?? {}}>
-                    <div class={clsx("placeholder", exiting() && "exiting")} />
-                </div>
-            </Show>
-        </div>
-    );
-};
